Adding implementation of the `howManyMales` method, updating contract from Integer return type to Long, this matches the `stream.count()` return type and prevents our solution being limited by the Integer.MAX_SIZE value.

+++ src/main/java/com/hopgood/alexander/PersonService.java
@@ -4,7 +4,7 @@ import com.hopgood.alexander.model.Person;
 
 public interface PersonService {
 
-    Integer howManyMales();
+    Long howManyMales();
 
     Person getOldest();
